Style: 로그인페이지 폴더구조 변경, 주석 정리
