--- src/app/(BeforeLogin)/login/_components/init.ts
@@ -1,2 +1,0 @@
-// 보일러플레이트용 임시 파일
-// 추후 이 파일은 지워주세요
